backend/requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
itsdangerous==2.2.0

# Validation & Settings
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic[email]==2.10.6

# Email
fastapi-mail==1.4.2
aiofiles==24.1.0

# Rate limiting
slowapi==0.1.9

# Redis (for WebSocket pub/sub)
redis==5.2.1

# Logging
loguru==0.7.3

# HTTP client (OAuth, testing)
httpx==0.28.1

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0

# File handling
pillow==11.1.0
python-magic==0.4.27

# Utilities
python-dateutil==2.9.0
